moving files to a new folder
